Renamed files Adafruit_sensor
This caused errors before
