+++ PiSky_Avionics/Pi_10DOF/Adafruit_LSM303/Adafruit_I2C.py
