+++ PiSky_Avionics/Pi_10DOF/__init__.py
@@ -1,5 +1,7 @@
 from L3GD20 import L3GD20
-from BMP085 import BMP085
+from Adafruit_BMP085 import BMP085
+from Adafruit_LSM303 import Adafruit_LSM303
+from Adafruit_I2C import Adafruit_I2C
 from LSM303 import LSM303
 import smbus
 import time
